+++ src/lib/pages/validator-details/components/validator-top/index.tsx
@@ -38,7 +38,6 @@ export const ValidatorTop = ({
       </Flex>
       <Flex direction="column" w="full" gap={{ base: 2, md: 1 }}>
         <ValidatorTitle info={info} />
-        {/* // TODO: Support LCD data */}
         <ValidatorStats
           validatorAddress={info.validatorAddress}
           commissionRate={info.commissionRate}